Setting version for the release 1.6.20

+++ tools/releaseScript.sh
@@ -64,7 +64,9 @@ git commit -a -m "Setting version for the release ${version}"
 printf "\n================================================================================\n"
 printf "Downloading Element Android source ${version}...\n"
 mkdir "./tmp/"
+if [ ! -f ./tmp/release.zip ]; then
 curl  https://github.com/vector-im/element-android/archive/refs/tags/v${version}.zip -i -L -o ./tmp/release.zip
+fi
 
 printf "\n================================================================================\n"
 printf "Unzipping Element Android source ${version}...\n"
